Add new API for Flatpak to CalculateSizeWithTask
This API is used from UI code just after new package selection is
resolved. This will allow us to react on package selection in Flatpak
immediately.

Also move CalculateFlatpaksSizeTask to initialization because this task
is not used for installation.

Related: RHEL-75804
(cherry picked from commit f4ca664c7ee5decb3d97f620783adbbafa8bcb81)
